Remove metrics after relaying to avoid memory growth
I have cases where sparsly used metrics cause the statsdpy process to
grow unbounded. The problem is that the array will keep tracking metrics
we are no longer seeing, causing it to grow over time. Eventually the
array is so big that it doesn't fit in memory, and accesing a particular
item out of the huge array is computationally expensive.

This patch simply deletes stuff from the tracking arrays after they are
relayed, instead of setting their values to 0 or empty array.